Retarget project to Unreal Engine 5.8
Pin EngineAssociation and Target.cs to 5.8 (V7 / Unreal5_8).
Use int32 for Blueprint-exposed StateFlags (UHT rejects uint32).

+++ Source/PurgatoriumLex/Combat/FighterCombatSim.cpp
@@ -63,7 +63,7 @@ uint32 FFighterCombatSim::ComputeStateHash() const
 	Mix(static_cast<uint32>(State.MovePhase));
 	Mix(static_cast<uint32>(State.AttackSnapshotPosture));
 	Mix(static_cast<uint32>(State.MoveLocalFrame));
-	Mix(State.StateFlags);
+	Mix(static_cast<uint32>(State.StateFlags));
 	Mix(static_cast<uint32>(State.Health));
 	return Hash;
 }
